+++ test/boost_test.dart
@@ -32,6 +32,8 @@ void main() {
     test('sortBy', _sortByTest);
     test('min / max', _minMaxTest);
     test('whereIs', _whereIsTest);
+    test('tuple a/b', _tupleIterableTest);
+    test('triple a/b/c', _tripleIterableTest);
   });
 
   group('Types', () {
@@ -321,6 +323,19 @@ void _whereIsTest() {
   expect(list.whereIs<bool>(), orderedEquals([true]));
   expect(list.whereIs<Null>(), orderedEquals([null]));
   expect(list.whereIs<double>(), orderedEquals([0.4]));
+}
+
+void _tupleIterableTest() {
+  final list = [Tuple('1', 1), Tuple('2', 2), Tuple('3', 3)];
+  expect(list.a, orderedEquals(['1', '2', '3']));
+  expect(list.b, orderedEquals([1, 2, 3]));
+}
+
+void _tripleIterableTest() {
+  final list = [Triple('1', 1, true), Triple('2', 2, 'abc'), Triple('3', 3, false)];
+  expect(list.a, orderedEquals(['1', '2', '3']));
+  expect(list.b, orderedEquals([1, 2, 3]));
+  expect(list.c, orderedEquals([true, 'abc', false]));
 }
 
 // Types